adjust the auto-text behavior of DrRacket so that it doesn't require the   queue-callback / execute callback dance; also, change the behavior   a little bit so that it works a little bit more like the rest of the   DrRacket languages; in particular, the initialization of the REPL   now only happens when a window is first opened or a new tab is first   created, but not at other times (ie not when the language changes;   when the language changes, we just keep the REPL state the same and   show a warning like before)
This change also required a change to the way the repl is initialized
and a slight change to the behavior of the first-opened method. Specifically,
it is now called in a slightly better context so that errors that
happen look like errors in the user's program. The only other use of
the first-opened method in the tree was to initialize the teachpacks
in the teaching languages and this new behavior is also an improvement
there.

original commit: bf53fd5c38e34f92bf55fb2b7ca871dfe2e2c3d3