clean up `set!' implementation in ASL to avoid catching failures
The code seems to pre-date the failure thunk to `syntax-local-value'.
A problem with letting an exception get raised is that it triggers
the `error-value->string' handler, which loads Slideshow pict support
in DrRacket, which loads `racket/draw' at phase 1 --- which isn't
inherenty bad, but a lot of work that can be avoided.

original commit: 7ddbc95b51538137d978c976b32e1332696958f2