make addition of `test~object` work with the new macro system
The old way of `(quasiquote (module .... ,#'(define ....)))` doesn't
work, because starting the module with an empty set of scopes means
that the added `define` has that empty set, and so it will end up
with an ambigious binding.

Injecting the `define` after expansion works, because the expanded
module has its own scopes that are not on the `define`.

An alternative would be to mark the original module source
using `make-syntax-introducer`, etc., but since there's already
a post-expansion adjustment to the module, moving the definition
there seems simplest.