Stepper remove special values mechanism
It appears that the before-after step comparison supersedes
special value detection. That is, we can observe that
(cons 3 empty) reduces to (cons 3 empty) and hide this step
even without knowing whether the applied function is the
built-in cons function. We can always bring this code
back if we need to...