part of fixing an error in error messages introduced in 4ad805c
I am not sure why this test suite passed before the error.
The error message contained <procedured:car> which is not
in the rewritten error message. I changed it to car, which
is what I am getting now. Removing the name would be bad
because that's precisely one of the things that went wrong
with the error message.