Repair broken PNG IDAT window lengths.
The "pngfix" utility detects problems in the four PlanetCuteShadow PNG
images. The output from the utility looks like,

  $ pngfix PlanetCuteShadow1.png
  iCCP TFB default 10 11 995 1960 PlanetCuteShadow1.png
  IDAT OK  default 15 15 57669 2072367 PlanetCuteShadow1.png

and the "TFB" line indicates a minor problem. On Gentoo, this is
outputting some harmless but scary QA warnings to end users.

Fortunately, pngfix is also able to correct the issue. This commit
updates those four images with the result of calling "pngfix -o" on
them.