Redo struct signatures
This is from a discussion Matthias Felleisen and I had at ICFP 2019:

The previous version of the signature support had as the name of the
signature the name of the struct, which made for confusing error
messages.

Now, camelcase the struct name: dillo -> Dillo, foo-bar -> FooBar.

This is in line with HtDP's conventions, I think.  Also, primitive
signatures start with an upper-case letter.

The signature constructors append an "Of": DilloOf, FooBarOf etc.