Killing the custodian that spawns the worlds and waits for their results is a bad thing for exceptions. It ends up triggering additional exceptions because the eventspaces get killed and the callbacks don't function.
Let's try this alternative solution:

- print the exception message
- re-raise the exception as before

The downside is that the universe continues to run and simply
records that one of the worlds has disappeared. In all likelihood
the same error will trip up other instances of the world.

A bit of experimentation with this alternative suggests drracket
reports errors in a way that students might be able to handle.