In menu-based *SL, don't print test-results twice for top-level exceptions
The menu-based *SL languages have a hack to make sure test results are
printed even if a top-level exception prevents the regular path to
printing.

The underlying machinery in the #lang languages has acquired a similar
hack, though, so we get the results printed twice in that case.

Remove the hack from the menu-based *SL languages.

Fixes #203.