Correctly count successful tests.
Fixes #212.

The test-engine previously assumed that, if there were no errors, the
number of successful tests would be equal to the total number of tests.

This is of course wrong when the tests didn't run.  As this can happen
now after the previous round of fixes in test execution, properly
count successful tests.

To that end, document that we're expecting test thunks to return
booleans.