Make big-bang stepper tests more robust
In big-bang tests, if the tick frequency is high, some to-draw steps
(e.g. the one immediately following the launch of big-bang) can be
absent in stepper because no such call are made.

This commit adds the `(repetition ,lo ,hi ,steps) form to stepper tests to
account for skippable steps.